backend/ocr/requirements.txt
google-cloud-vision==3.4.1
google-cloud-storage==2.10.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
